requirements.txt
Django==1.10.6
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.3
olefile==0.44
Pillow==4.0.0
